Issue #20419: Add test to ensure xpoweredby can't be enabled..
