+++ dev/io.openliberty.webcontainer.servlet.6.0.internal_fat/bnd.bnd
@@ -35,5 +35,7 @@ javac.target: 11
 # For all project names that match the pattern "*_fat*", dependencies for junit,
 # fattest.simplicity, and componenttest-1.0 will be automatically added to the buildpath
 -buildpath: \
-    io.openliberty.jakarta.servlet.6.0;version=latest
+    io.openliberty.jakarta.servlet.6.0;version=latest,\
+    org.apache.httpcomponents.client5:httpclient5;version=5.0.2,\
+    org.apache.httpcomponents.core5:httpcore5;version=5.0.2
