+++ dev/io.openliberty.webcontainer.servlet.6.0.internal_fat/fat/src/io/openliberty/webcontainer/servlet60/fat/FATSuite.java
@@ -14,11 +14,11 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
 
-import io.openliberty.webcontainer.servlet60.fat.tests.SimpleTest;
+import io.openliberty.webcontainer.servlet60.fat.tests.Servlet60XPoweredByHeaderTest;
 
 @RunWith(Suite.class)
 @SuiteClasses({
-                SimpleTest.class
+                Servlet60XPoweredByHeaderTest.class
 })
 public class FATSuite {
