+++ dev/io.openliberty.webcontainer.servlet.6.0.internal_fat/publish/servers/servlet60_xPoweredByHeaderTest/bootstrap.properties
