+++ dev/io.openliberty.webcontainer.servlet.6.0.internal_fat/publish/servers/servlet60_xPoweredByHeaderTest/server.xml
@@ -8,12 +8,15 @@
     Contributors:
         IBM Corporation - initial API and implementation
  -->
-<server description="Server for testing Webcontainer">
+<server description="Server for test disableXPoweredBy configuration.r">
 
     <include location="../fatTestPorts.xml"/>
 
     <featureManager>
         <feature>servlet-6.0</feature>
     </featureManager>
+    
+     <webContainer disableXPoweredBy="false"/>
+    
 
 </server>
